refactor dedicated code into consts and types
