+++ src/dev/project/postgres_data
@@ -1,0 +1,1 @@
+/home/user/postgres_data/
