+++ src/packages/util/consts/billing.ts
@@ -1,0 +1,9 @@
+/*
+ *  This file is part of CoCalc: Copyright © 2020 Sagemath, Inc.
+ *  License: AGPLv3 s.t. "Commons Clause" – see LICENSE.md for details
+ */
+
+export const ONE_DAY_MS = 24 * 60 * 60 * 1000;
+export const AVG_MONTH_DAYS = 30.5;
+export const AVG_YEAR_DAYS = 12 * AVG_MONTH_DAYS;
+export const ONE_MONTH_MS = AVG_MONTH_DAYS * ONE_DAY_MS;
