+++ src/packages/util/db-schema/site-licenses.ts
@@ -20,61 +20,15 @@ for students).
 import { is_valid_uuid_string, plural } from "../misc";
 import { Table } from "./types";
 import { SCHEMA } from "./index";
-
-export const ONE_DAY_MS = 24 * 60 * 60 * 1000;
-export const AVG_MONTH_DAYS = 30.5;
-export const AVG_YEAR_DAYS = 12 * AVG_MONTH_DAYS;
-export const ONE_MONTH_MS = AVG_MONTH_DAYS * ONE_DAY_MS;
-
-export type DedicatedDiskTypes = "ssd" | "standard" | "balanced";
-
-export type DedicatedDisk =
-  | {
-      size_gb: number;
-      type: DedicatedDiskTypes;
-      name: string; // the ID of the disk, globally unique, derived from the license-ID, generated upon license creation or maybe manually.
-    }
-  | false;
-
-export type DedicatedVM = {
-  name: string; // the ID of the VM, globally unique, derived from the license-ID, generated upon license creation or maybe manually.
-  machine: string;
-};
-
-export const DISK_NAMES: { [type in DedicatedDiskTypes]: string } = {
-  standard: "slow",
-  balanced: "medium",
-  ssd: "fast",
-};
-
-export function isDedicatedDisk(d): d is DedicatedDisk {
-  return (
-    d != null &&
-    typeof d.size_gb === "number" &&
-    ["ssd", "standard", "balanced"].includes(d.type)
-  );
-}
+import {
+  DedicatedDisk,
+  DedicatedVM,
+  DISK_NAMES,
+} from "@cocalc/util/types/dedicated";
 
 export function dedicated_disk_display(disk: DedicatedDisk): string {
   if (typeof disk === "boolean") return "";
   return `${disk.size_gb} GiB, ${DISK_NAMES[disk.type] ?? disk.type} speed`;
-}
-
-export interface VMsType {
-  [id: string]: {
-    name?: string;
-    price_day: number;
-    spec: { mem: number; cpu: number };
-    quota: { dedicated_vm: string }; // only those defined in VMS below
-  };
-}
-
-export interface DiskType {
-  [id: string]: {
-    name: string;
-    price_day: number;
-    quota: { dedicated_disk: { size_gb: number; type: DedicatedDiskTypes } };
-  };
 }
 
 export interface Quota {
